Completed

+++ app/src/main/java/com/example/sagar/day1pro/MainActivity.java
@@ -11,6 +11,7 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ProgressBar;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import java.io.BufferedInputStream;
@@ -32,7 +33,9 @@ public class MainActivity extends AppCompatActivity {
     String pdfLink;
     int downSize=0;
     ProgressBar progressBar;
+    TextView progressPercent;
     File file;
+    boolean check=true;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -53,22 +56,26 @@ public class MainActivity extends AppCompatActivity {
     {
         AlertDialog.Builder dialog=new AlertDialog.Builder(this);
         dialog.setTitle("Download");
-        dialog.setMessage("Choose the Option");
+        dialog.setMessage("Choose the Option").setCancelable(false);
         LayoutInflater layoutInflater=LayoutInflater.from(this);
         final View popup_layout=layoutInflater.inflate(R.layout.popup_layout,null);
         progressBar=popup_layout.findViewById(R.id.progressBar);
-        dialog.setView(popup_layout);
-        dialog.setPositiveButton("Download", new DialogInterface.OnClickListener() {
+        progressPercent=popup_layout.findViewById(R.id.progressPercent);
+        progressPercent.setText("0%");
+        dialog.setCancelable(false).setPositiveButton("Download", new DialogInterface.OnClickListener() {
             @Override
             public void onClick(DialogInterface dialog, int which)
             {
-                downloadPdf();
+
             }
         });
         dialog.setNegativeButton("Stop Download", new DialogInterface.OnClickListener() {
             @Override
-            public void onClick(DialogInterface dialog, int which) {
-                if(file.exists())
+            public void onClick(DialogInterface dialog, int which)
+            {
+                Toast.makeText(MainActivity.this, "Download Cancelled", Toast.LENGTH_SHORT).show();
+                check=false;
+                if(file!=null&&file.exists())
                 {
                     file.delete();
                 }
@@ -77,15 +84,25 @@ public class MainActivity extends AppCompatActivity {
         });
         dialog.setNeutralButton("Do In Background", new DialogInterface.OnClickListener() {
             @Override
-            public void onClick(DialogInterface dialog, int which) {
+            public void onClick(DialogInterface dialog, int which)
+            {
+                check=true;
                 return;
             }
         });
-        dialog.show();
+        dialog.setView(popup_layout);
+        AlertDialog dialog1=dialog.create();
+        dialog1.show();
+        dialog1.getButton(DialogInterface.BUTTON_POSITIVE).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                check=true;
+                downloadPdf();
+            }
+        });
     }
     private void downloadPdf()
     {
-        Toast.makeText(this, pdfLink, Toast.LENGTH_SHORT).show();
         Call<ResponseBody> call = ApiClient.getInstance().getPdfAPI().getPdf(pdfLink);
         call.enqueue(new Callback<ResponseBody>() {
             @Override
@@ -104,11 +121,9 @@ public class MainActivity extends AppCompatActivity {
 
     private void writeResponseBodyToTheDisk(ResponseBody body)
     {
-        Toast.makeText(MainActivity.this, "Reached Response", Toast.LENGTH_SHORT).show();
         OutputStream outputStream = null;
         InputStream inputStream=new BufferedInputStream(body.byteStream());
         try {
-
             file = new File(this.getFilesDir(),"Hello");
 
             // write the inputStream to a FileOutputStream
@@ -118,18 +133,23 @@ public class MainActivity extends AppCompatActivity {
             int read = 0;
             byte[] bytes = new byte[1024];
 
-            while ((read = inputStream.read(bytes)) != -1) {
+            while (((read = inputStream.read(bytes)) != -1)&& check) {
                 downSize += read;
                 //Log.e("tag1",""+total);
                 //Log.e("tag2",""+((total*100)/2801192));
                 outputStream.write(bytes, 0, read);
             }
-            Toast.makeText(MainActivity.this, (file.toString())+"", Toast.LENGTH_SHORT).show();
-            Toast.makeText(MainActivity.this, file.getPath()+"", Toast.LENGTH_LONG).show();
+            if(!check)
+            {
+                if(file!=null && file.exists())
+                {
+                    file.delete();
+                }
+                return;
+            }
             downloadComplete(pdfLink);
         } catch (IOException e) {
             e.printStackTrace();
-            Toast.makeText(MainActivity.this, "Error try occured", Toast.LENGTH_SHORT).show();
         } finally {
             if (outputStream != null) {
                 try {
@@ -151,6 +171,7 @@ public class MainActivity extends AppCompatActivity {
         dialog.setMessage("Choose the Option");
         LayoutInflater layoutInflater=LayoutInflater.from(this);
         final View popup_layout=layoutInflater.inflate(R.layout.popup_layout,null);
+        progressPercent.setText("100%");
         dialog.setView(popup_layout);
         dialog.setPositiveButton("Open Pdf", new DialogInterface.OnClickListener() {
             @Override
